Refactor publish workflow to be generic (#96)
Right now, adding a new library involves a lot of copy/pasting of code in publish.yml. It also is likely to cause errors as the order in which the packages are published matters. This commit makes the build order very apparent, and then uses the same steps to properly publish the packages

Implements #95