Properly convert CHANGED_PACKAGES values to bools (#104)
Dorny/filter returns a map where the values are strings. This fixes the logic to be booleans, so the rest of the publish detection works.

There is an additional bug where the quotes were getting lost in the set output step, due to the bash parsing of the variable. Adding single quotes preserves the double quotes within the string